+++ voting-circuits/src/circuit/mod.rs
@@ -4,6 +4,7 @@ pub(crate) mod address_ownership;
 pub(crate) mod elgamal;
 pub(crate) mod gadget;
 pub(crate) mod mul_chip;
+pub(crate) mod nonzero;
 pub(crate) mod poseidon_merkle;
 pub(crate) mod spend_authority;
 pub(crate) mod van_integrity;